refactor: add cache for sync block mark reader
