+++ src/query/storages/common/cache/src/lib.rs
@@ -40,5 +40,6 @@ pub use read::InMemoryBytesCacheReader;
 pub use read::InMemoryItemCacheReader;
 pub use read::LoadParams;
 pub use read::Loader;
+pub use read::SyncLoader;
 
 pub use self::metrics::*;
