+++ src/query/storages/common/cache/src/read/mod.rs
@@ -21,5 +21,6 @@ pub use cached_reader::CachedReader;
 pub use loader::CacheKey;
 pub use loader::LoadParams;
 pub use loader::Loader;
+pub use loader::SyncLoader;
 pub use readers::InMemoryBytesCacheReader;
 pub use readers::InMemoryItemCacheReader;
