+++ src/query/storages/fuse/src/io/read/delete/delete_mark_reader_sync.rs
@@ -16,27 +16,22 @@
 use std::sync::Arc;
 
 use common_arrow::arrow::bitmap::Bitmap;
-use common_arrow::arrow::datatypes::DataType;
-use common_arrow::arrow::datatypes::Field as ArrowField;
-use common_arrow::arrow::io::parquet::read::column_iter_to_arrays;
-use common_arrow::parquet::compression::Compression;
 use common_arrow::parquet::metadata::ColumnChunkMetaData;
-use common_arrow::parquet::read::BasicDecompressor;
-use common_arrow::parquet::read::PageMetaData;
-use common_arrow::parquet::read::PageReader;
 use common_exception::ErrorCode;
 use common_exception::Result;
-use common_expression::types::BooleanType;
-use common_expression::types::ValueType;
-use common_expression::Column;
 use opendal::Operator;
 use storages_common_cache::CacheKey;
+use storages_common_cache::CachedReader;
 use storages_common_cache::LoadParams;
+use storages_common_cache::SyncLoader;
+use storages_common_cache_manager::CachedObject;
 use storages_common_cache_manager::DeleteMarkMeta;
 use storages_common_table_meta::meta::DeleteMaskVersion;
 use storages_common_table_meta::meta::Location;
 
-use crate::io::read::delete::delete_mark_reader::DeleteMetaLoader;
+use crate::io::read::delete::delete_mark_reader::deserialize_bitmap;
+use crate::io::read::delete::delete_mark_reader_async::DeleteMetaLoader;
+use crate::io::MetaReaders;
 
 // TODO 1. code duplication 2. cache
 pub trait DeleteMarkSyncReader {
@@ -99,44 +94,33 @@ fn sync_load_delete_mark_data<'a>(
         column_descriptor: col_chunk_meta.descriptor().clone(),
     };
 
-    // let cached_reader = CachedReader::new(Bitmap::cache(), loader);
+    let cached_reader = CachedReader::new(Bitmap::cache(), loader);
 
     let param = LoadParams {
         location,
         len_hint: None,
         ver: 0,
     };
-    // let bytes = cached_reader.read(&param)?;
-    let bytes = loader.sync_load(&param)?;
-    Ok(Arc::new(bytes))
+    cached_reader.sync_read(&param)
 }
 
 /// Loads index meta data
 /// read data from cache, or populate cache items if possible
 #[tracing::instrument(level = "debug", skip_all)]
 fn sync_load_mark_meta(dal: Operator, path: &str, size: u64) -> Result<Arc<DeleteMarkMeta>> {
-    // TODO cache
-    let object = dal.object(path);
-    let mut reader = object.blocking_range_reader(0..size)?;
+    let path_owned = path.to_owned();
+    let reader = MetaReaders::delete_mark_meta_reader(dal);
+    // Format of FileMetaData is not versioned, version argument is ignored by the underlying reader,
+    // so we just pass a zero to reader
+    let version = 0;
 
-    use common_arrow::parquet::read::read_metadata;
-    let meta = read_metadata(&mut reader).map_err(|err| {
-        ErrorCode::Internal(format!("read file meta failed, {}, {:?}", path, err))
-    })?;
+    let load_params = LoadParams {
+        location: path_owned,
+        len_hint: Some(size),
+        ver: version,
+    };
 
-    Ok(Arc::new(DeleteMarkMeta(meta)))
-}
-
-/// Loads an object from storage
-#[async_trait::async_trait]
-pub trait SyncLoader<T> {
-    /// Loads object of type T, located by [params][LoadParams].
-    fn sync_load(&self, params: &LoadParams) -> Result<T>;
-
-    /// the [CacheKey] returns will be used as the key of cached item.
-    fn cache_key(&self, params: &LoadParams) -> CacheKey {
-        params.location.clone()
-    }
+    reader.sync_read(&load_params)
 }
 
 impl SyncLoader<Bitmap> for DeleteMetaLoader {
@@ -144,49 +128,7 @@ impl SyncLoader<Bitmap> for DeleteMetaLoader {
         let reader = self.operator.object(&params.location);
         let bytes = reader.blocking_range_read(self.offset..self.offset + self.len)?;
 
-        let page_meta_data = PageMetaData {
-            column_start: 0,
-            num_values: self.num_rows as i64,
-            compression: Compression::Uncompressed,
-            descriptor: self.column_descriptor.descriptor.clone(),
-        };
-
-        let page_reader = PageReader::new_with_page_meta(
-            std::io::Cursor::new(bytes), /* we can not use &[u8] as Reader here, lifetime not valid */
-            page_meta_data,
-            Arc::new(|_, _| true),
-            vec![],
-            usize::MAX,
-        );
-
-        let decompressor = BasicDecompressor::new(page_reader, vec![]);
-        let column_type = self.column_descriptor.descriptor.primitive_type.clone();
-        let filed_name = self.column_descriptor.path_in_schema[0].to_owned();
-        let field = ArrowField::new(filed_name, DataType::Boolean, false);
-        let mut array_iter = column_iter_to_arrays(
-            vec![decompressor],
-            vec![&column_type],
-            field,
-            None,
-            self.num_rows,
-        )?;
-        if let Some(array) = array_iter.next() {
-            let array = array?;
-            let col =
-                Column::from_arrow(array.as_ref(), &common_expression::types::DataType::Boolean);
-
-            let bitmap = BooleanType::try_downcast_column(&col).ok_or_else(|| {
-                ErrorCode::Internal(
-                    "unexpected exception: load delete mark raw data as boolean failed",
-                )
-            })?;
-
-            Ok(bitmap)
-        } else {
-            Err(ErrorCode::StorageOther(
-                "delete mark data not available as expected",
-            ))
-        }
+        deserialize_bitmap(&bytes, self.num_rows, &self.column_descriptor)
     }
 
     fn cache_key(&self, _params: &LoadParams) -> CacheKey {
